+++ scripts/dzVents/runtime/dzVents.lua
@@ -26,8 +26,8 @@ end
 
 -- todo remove!!!!!!
 --print('!!!!!!!!!! remove')print('!!!!!!!!!! remove')
-local persistence = require('persistence')
-persistence.store(currentPath .. '/domoticzData.lua', domoticzData)
+--local persistence = require('persistence')
+--persistence.store(currentPath .. '/domoticzData.lua', domoticzData)
 commandArray = {}
 
 utils.log('dzVents version: 2.2.1', utils.LOG_DEBUG)